+++ sdk/storage/azure_storage_blob/src/clients/block_blob_client.rs
@@ -38,12 +38,12 @@ impl GeneratedBlockBlobClient {
     ///
     /// # Arguments
     ///
-    /// * `block_blob_url` - The full URL of the block blob, for example `https://myaccount.blob.core.windows.net/mycontainer/myblob`.
+    /// * `blob_url` - The full URL of the block blob, for example `https://myaccount.blob.core.windows.net/mycontainer/myblob`.
     /// * `credential` - An optional implementation of [`TokenCredential`] that can provide an Entra ID token to use when authenticating.
     /// * `options` - Optional configuration for the client.
     #[tracing::new("Storage.Blob.BlockBlob")]
     pub fn from_url(
-        block_blob_url: Url,
+        blob_url: Url,
         credential: Option<Arc<dyn TokenCredential>>,
         options: Option<BlockBlobClientOptions>,
     ) -> Result<Self> {
@@ -56,10 +56,10 @@ impl GeneratedBlockBlobClient {
             .push(storage_headers_policy);
 
         let per_retry_policies = if let Some(token_credential) = credential {
-            if !block_blob_url.scheme().starts_with("https") {
+            if !blob_url.scheme().starts_with("https") {
                 return Err(azure_core::Error::with_message(
                     azure_core::error::ErrorKind::Other,
-                    format!("{block_blob_url} must use https"),
+                    format!("{blob_url} must use https"),
                 ));
             }
             let auth_policy: Arc<dyn Policy> = Arc::new(BearerTokenCredentialPolicy::new(
@@ -81,7 +81,7 @@ impl GeneratedBlockBlobClient {
         );
 
         Ok(Self {
-            endpoint: block_blob_url,
+            endpoint: blob_url,
             version: options.version,
             pipeline,
         })